requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async and HTTP
httpx>=0.25.0
aiohttp>=3.9.0

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0
cohere>=4.47

# Local models
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Vector database and embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.0
numpy>=1.24.0

# Clustering and visualization
scikit-learn>=1.3.0
umap-learn>=0.5.5
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Dashboard
streamlit>=1.31.0

# Graph analysis
networkx>=3.2.1
python-louvain>=0.16

# Data processing
pandas>=2.0.0
scipy>=1.11.0

# Utilities
python-dateutil>=2.8.2
rich>=13.7.0
loguru>=0.7.2
tqdm>=4.66.0
pyyaml>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Safety and security
python-jose>=3.3.0
cryptography>=42.0.0
